Move dynamic event code to its own module
